+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/checkers/DependencyGraphChecker.kt
@@ -3,11 +3,14 @@
 package dev.zacsweers.metro.compiler.fir.checkers
 
 import dev.zacsweers.metro.compiler.fir.FirMetroErrors
+import dev.zacsweers.metro.compiler.fir.additionalScopesArgument
 import dev.zacsweers.metro.compiler.fir.allAnnotations
 import dev.zacsweers.metro.compiler.fir.annotationsIn
 import dev.zacsweers.metro.compiler.fir.classIds
 import dev.zacsweers.metro.compiler.fir.findInjectConstructors
 import dev.zacsweers.metro.compiler.fir.isAnnotatedWithAny
+import dev.zacsweers.metro.compiler.fir.resolvedAdditionalScopesClassIds
+import dev.zacsweers.metro.compiler.fir.resolvedScopeClassId
 import dev.zacsweers.metro.compiler.fir.scopeAnnotations
 import dev.zacsweers.metro.compiler.fir.validateApiDeclaration
 import org.jetbrains.kotlin.diagnostics.DiagnosticReporter
@@ -26,6 +29,7 @@ import org.jetbrains.kotlin.fir.resolve.firClassLike
 import org.jetbrains.kotlin.fir.types.coneTypeOrNull
 import org.jetbrains.kotlin.fir.types.isNothing
 import org.jetbrains.kotlin.fir.types.isUnit
+import org.jetbrains.kotlin.name.StandardClassIds
 
 // TODO
 //  - if there's a factory(): Graph in the companion object, error because we'll generate it
@@ -63,6 +67,19 @@ internal object DependencyGraphChecker : FirClassChecker(MppCheckerKind.Common) 
         )
         return
       }
+    }
+
+    // Ensure scope is defined if any additionalScopes are defined
+    val scope =
+      dependencyGraphAnno.resolvedScopeClassId()?.takeUnless { it == StandardClassIds.Nothing }
+    val additionalScopes = dependencyGraphAnno.resolvedAdditionalScopesClassIds().orEmpty()
+    if (additionalScopes.isNotEmpty() && scope == null) {
+      reporter.reportOn(
+        dependencyGraphAnno.additionalScopesArgument()?.source ?: dependencyGraphAnno.source,
+        FirMetroErrors.DEPENDENCY_GRAPH_ERROR,
+        "@${graphAnnotationClassId.shortClassName.asString()} should have a primary `scope` defined if `additionalScopes` are defined.",
+        context,
+      )
     }
 
     declaration.validateApiDeclaration(